restructuration on going
