+++ pages/login.jsx
@@ -2,7 +2,7 @@ import{Container, Button, TextField} from '@mui/material';
 import {useState, useContext} from 'react';
 import { appContext } from './context/appContext';
 import Swal from'sweetalert2';
-import {useRouter} from 'next/router';
+import {makePublicRouterInstance, useRouter} from 'next/router';
 
 function Login(){
 
@@ -17,6 +17,10 @@ function Login(){
 
     const handleChange = (e) =>{
         setRegisterForm({...registerForm, [e.target.id]: [e.target.value]});
+    }
+
+    const goCreateAccount = ()=>{
+        router.push('/register');
     }
 
     const errorModal = ()=>{
@@ -82,6 +86,7 @@ function Login(){
             </form>
             <Button onClick={fireLogout}>Logout</Button>
             <Button onClick={goToPrivatePage}>Go to privatePage</Button>
+            <Button onClick={goCreateAccount}>Create account</Button>
         </Container>
         </>
     )
